chill

+++ src/rs/lib.rs
@@ -8,7 +8,7 @@ extern "C" {
 }
 
 #[wasm_bindgen]
-pub struct Point3f {
+pub struct Point3d {
     pub x: f64,
     pub y: f64,
     pub z: f64
@@ -26,8 +26,8 @@ pub fn why(x: &str) -> f64 {
 }
 
 #[wasm_bindgen]
-pub fn project(x: f64, y: f64, z: f64, a: f64, b: f64, c: f64, d: f64, e: f64, f: f64) -> Point3f {
-    Point3f { x: x + y + z, y: 2.0, z: 3.0 }
+pub fn project(x: f64, y: f64, z: f64, a: f64, b: f64, c: f64, d: f64, e: f64, f: f64) -> Point3d {
+    Point3d { x: x + y + z, y: 2.0, z: 3.0 }
 }
 
 #[cfg(test)]
